updates generated paging and observable classes

+++ Octokit.Reactive/Clients/IObservableDeploymentStatusClient.cs
@@ -21,7 +21,7 @@ namespace Octokit.Reactive
         /// <param name="owner">The owner of the repository.</param>
         /// <param name="name">The name of the repository.</param>
         /// <param name="deploymentId">The id of the deployment.</param>
-        IObservable<DeploymentStatus> GetAll(string owner, string name, int deploymentId);
+        IObservable<DeploymentStatus> GetAll(string owner, string name, long deploymentId);
 
         /// <summary>
         /// Gets all the statuses for the given deployment. Any user with pull access to a repository can
@@ -32,7 +32,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="deploymentId">The id of the deployment.</param>
-        IObservable<DeploymentStatus> GetAll(long repositoryId, int deploymentId);
+        IObservable<DeploymentStatus> GetAll(long repositoryId, long deploymentId);
 
         /// <summary>
         /// Gets all the statuses for the given deployment. Any user with pull access to a repository can
@@ -45,7 +45,7 @@ namespace Octokit.Reactive
         /// <param name="name">The name of the repository.</param>
         /// <param name="deploymentId">The id of the deployment.</param>
         /// <param name="options">Options for changing the API response</param>
-        IObservable<DeploymentStatus> GetAll(string owner, string name, int deploymentId, ApiOptions options);
+        IObservable<DeploymentStatus> GetAll(string owner, string name, long deploymentId, ApiOptions options);
 
         /// <summary>
         /// Gets all the statuses for the given deployment. Any user with pull access to a repository can
@@ -57,7 +57,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="deploymentId">The id of the deployment.</param>
         /// <param name="options">Options for changing the API response</param>
-        IObservable<DeploymentStatus> GetAll(long repositoryId, int deploymentId, ApiOptions options);
+        IObservable<DeploymentStatus> GetAll(long repositoryId, long deploymentId, ApiOptions options);
 
         /// <summary>
         /// Creates a new status for the given deployment. Users with push access can create deployment
@@ -70,7 +70,7 @@ namespace Octokit.Reactive
         /// <param name="name">The name of the repository.</param>
         /// <param name="deploymentId">The id of the deployment.</param>
         /// <param name="newDeploymentStatus">The new deployment status to create.</param>
-        IObservable<DeploymentStatus> Create(string owner, string name, int deploymentId, NewDeploymentStatus newDeploymentStatus);
+        IObservable<DeploymentStatus> Create(string owner, string name, long deploymentId, NewDeploymentStatus newDeploymentStatus);
 
         /// <summary>
         /// Creates a new status for the given deployment. Users with push access can create deployment
@@ -82,6 +82,6 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="deploymentId">The id of the deployment.</param>
         /// <param name="newDeploymentStatus">The new deployment status to create.</param>
-        IObservable<DeploymentStatus> Create(long repositoryId, int deploymentId, NewDeploymentStatus newDeploymentStatus);
+        IObservable<DeploymentStatus> Create(long repositoryId, long deploymentId, NewDeploymentStatus newDeploymentStatus);
     }
 }
